Create 206. Reverse Linked List
I know I have to keep one pointer going through the list. I also have to update the nearest node next to dummy because its next is not the next in list therefore I will have to update the pointer as well.

Dummy node is good for singly linked list.