Create 437 Path Sum III
I was stuck on how to find the right path in the middle of the tree. I know how to find the results from the root, which can be obtained by performing dfs. 

However, if you just use recursion on the function itself, it can traverse the tree again.

I think it's the first time that I ever encounter a problem that requires traversals in two functions.