Create 102. Binary Tree Level Order Traversal
I learnt a new trick for BFS of trees: for level order traversals, use two containers, one for current level and one for next level. The container can be queues, stacks, vectors, depending on what the question is asking. After you are done, swap the two containers.

The use of pointers: if you swap the two containers by values, it will be really difficult. However, if you just swap pointers, it will be easy.

Couple mistakes I made:

When I created cur and next pointers, initially I forgot to put new statements, which led to run time errors, dumb!

when I swap the containers, I have both cur and next point to the next level containers, which led to the always empty container, therefore only the very first level is traversed.

This is almost like a template, won't be a bad idea to get it down well.