Create 96. Unique Binary Search Trees
Such a short program, yet it's the first dp problem I have done, and I struggled with it.

1. for the non-recursive solution, the time complexity is O(n^2) for the double for loop. Because for the outer loop, you have to build n = 2, n = 3, and all the way up to n. We know two trivial situations, when n = 0, and n = 1. The inner loop is to use the pattern:

f(n) = f(0)*f(n-1) + f(1)*f(n-2)+ ... + f(n-1)*f(0).

2. recursive solution

now we have seen several cases like this: representing a tree using an array, vector. The method is to use a low-high index combo to shrink the scope of the problem. The base case is always when the low index is == high index, leaf node; when the low index > high index, NULL.