Create 222. Count Complete Tree Nodes
This problem uses the fact that a full binary tree with height h where h starts from 1 is 2^h -1. It then recursively checked if the current tree is a full binary tree. If yes, then return using the formula. If not, then continues. It will stop because the base case can be a leaf node. which will return 1. 

Another trick is using 2<<h to replace pow(2, h+1). When h = 0, it will be 2.