Create 450. Delete Node in a BST
This problem cost me a day. No kidding.

Looking back, this problem use recursion for backtracking and build the tree from bottom up and can be divided into these subproblems:

1. base case: as always.if null return null.

2. if found(delete node, maintain structure)

3. root->left =; root->right = ; 

4. return root;

what makes this problem hard is that in #2 there is a lot.

1. leaf node, easy return null. then the father node links to a null pointer, gone and done.

2. has one child, you have to then figure out which child it has and then return it.

3. two children. You have to find the predecessor, return this predecessor after you link it to the children of the node you want to delete. and when you are finding the predecessor, you have to remove it from the tree.

What makes 3 so messy is that there are a lot of scenarios to be taken care of: what if the predecessor is the right child. what if the predecessor has subtrees that need be linked to the father of the predecessor.

another solution is to use build tree to rebuild the tree without the node that is deleted. In order traversal to get the sorted vector, use the vector again to build the tree you want.