Create 337. House Robber III
I approached this question by top-down method and it's a mess. 

But for bottom up. Imagine there's three levels. All you need to do is to find out the results from next-next level, and then compare who is larger, next-next-level + root, or just next level. That's all method 1 is about. However, there's a lot of repetition. There's overlap in the recursive calls and It will be great if we can store the information. In fact, the first method will have a time out error.

That will lead us to the second method, which is quite neat. You will use a two-element array, which marks the bi-state situation.