Create 058. Length of Last Word
You have to be able to see all the possible situations and you have to abstract it in the most simple way. 

Here is the situation:

^^^^^^^^^###^^^^^^

^ is space and # is char. Therefore instead of considering a lot of other situations, this is the name of the game.

Our job here is to count of course, but where do we start counting, and when do we return and how do we take care of the continuous situations. 

My approach is to have a state marker that flips on when it sees the first char. Therefore it will just skip all the continuous space chars.

My previous mistake is that I didn't flip on the bool when I see a non-space char. Therefore it will keep going.