Create 088. Merge Sorted Array
If you start from the front, you will see that you have to shift all the elements if you happen to have an element from the second array that is smaller than the element pointed in the first array. What to do, you start from the back.

Takeaway:

remember the continue.