Create 236. Lowest Common Ancestor of a Binary Tree
First solution is just like a binary search tree but it has a time out error.

The correct solution still employs the same logic: if the root is one of the two elements, then return the root because of the hierarchy: nothing beats the root.

Then what you want to do is to find out if the two are in:

1. in both left and right SUBTREES

2. both in left SUBTREES

3. both in right SUBTREES

The function then turns into a finder function: return the element same as p or q. The reason I am highlighting the subtrees is that I made a mistake earlier using root->left for return values. This is absurd. Because there is no way to represent result down below when you just return the left and right child of the root.

Another lesson learnt is how to return value from deep down the tree: using the return value. When you hit a base case, you will return root, then the root is trickled up to a right or left, and keep being returned until it hits the root for being returned for the last time.

In 20/20 hindsight, this problem is easy, lol.