Create target path
This is one part of the solution of LCA from geeksforgeeks that I found interesting. It returns a path that contains a target node in a tree.

If I were to do it, I will create I turn series: a vector of bools to signal how I get here, a true means coming from right and a false means the left of a root. Then I just go from the root, following the directions and reach the target node.

But this solution is neat. It uses a bool to stop the recursion when it hits the target node, then it also uses it to signal whether this node should be in the final result or not. It not, it just pops it out, but it always puts it in the final result first.