Create 136. Single Number
The first method is straightforward but requires O(n) extra memory. This also oks _map[nums[i]] = 0 if it's the first time found in map.

The second solution uses the following method:

a xor 0 = a; a xor a = 0

Therefore xor every element with 0 will cancel out the even-number appearance and will leave the element with only one appearance.

And you can move around the elements with xor.