Create 137. Single Number II
solution 1 is just like single number 1.

solution 2 uses this fact: if all numbers appear three times, then the sum of a certain bit mod 3 will be 0. Because if the digit is 0, that is obviously right. If a digit is 1, then since it appears three times, it must be 0.

Therefore solution 2 uses the following logic: for a 32-bit number, we go digit by digit to build this number by finding what digit is 1 and bit or it with 0. 

How to determine which bit is not 0? Use the mod logic: for each of the 32 bit number, right shift 1 by i(i is from 0 to 31). Then for all the numbers, if the number & the mask(with only 1 and others 0) is 1, that means that digit of the scanned number is 1. Then you mod the count with 3 to figure out if you want to create a '1' in the final number.

This is my first experience with bit manipulation.