Create 409. Longest Palindrome
This problem is easier than I thought. Originally took me a bit of detour to get to the right point. 

Clear logic: palindrome starts from the middle and rippled to the ends. Therefore, if a char appears even number of time, we can just place them symmetrically, so we just add them to the final result. For odd number appearance, if it's the first time that we have an odd, we can have one in the middle and then place the rest symmetrically. No wastes. But if the center position has been filled, we just have to throw one away.

I used a bool to record if the center pos has been filled. I forgot to flip it after I filled it, which of course is an error.