Create 447. Number of Boomerangs
This problem takes me a while to think and a while to debug.

It's really pretty straightforward in 20/20 hindsight. We will just do it point by point. For each point, create a hashtable to record the number of equal distance, then use it to create permutations.

The mistake I made is in the second for loop:

I did: for(int j = i+1; j < points.size(); j++)

This, of course, has been done many times. but it's wrong here.

Always write down your logic on a paper before you code. It is a good practice.