Create 242. Valid Anagram
The first solution is just easy and elegant but I never thought of this question and this is actually the first problem that I use sort. Not a bad idea to be thrown into the mix. 

Second solution uses a hashtable to solve the problem and it uses chars as the key which also can be transformed into indices in arrays  thus avoid figuring out what is exactly inside the hashtable and the counts.

The third solution does use the hashtable and you don't have to figure out what is inside the hashtable. 

I do not know what's inside a hash table I don't know how to untangle what the keys are and what the counts are. Therefore using these two methods I don't have to know what's inside the blackbox I will just use the relationship provided by the problem. to solve the problem itself.