Create 003. Longest Substring Without Repeating Characters
I almost had it. But dvdf is a counter. My solution doesn't account for moving from the second element to the last element. 

Now the hint shows that I can use hashtable to solve this problem, which I have no idea to do it.

Now I understand why I should use hashtable. Instead of using isduplicate, which takes O(n) and need to maintain an extra string. Use a  vector<int> to determine if the char has appeared before or not.