Create 349. Intersection of Two Arrays
The reason why it works is that set doesn't allow duplicates. Therefore when you construct the set, you don't have to worry about duplicates. But, when you go through the second vector, you have to erase the elements from the set because you have to keep duplicates from the final result.