Create Remove Duplicates from Sorted Array
I use the right method for sorted array: two pointers. 

The right way to do this is use two pointers: one scan the whole array and one to mark the end of the correct part. Now when you meet duplicates, just pass it; when you find different values, swap it to the end of the right part, and increment the "boundary" if you will, and return the boundary index  + 1 because it's asking for size.