Create 121. Best Time to Buy and Sell Stock
This is a good example of limited scope thinking. I was stuck on just using two pointers to go through the array because I just did a problem on that. In fact, if I just finished doing the maximum subarray, I will more than likely to use dynamic programming, just like I should be doing in this problem.