Create 229. Majority Element II
This problem builds on the first part. The logic still the same. However, in this case, the sequence matters. 

if you switch the if statements in the first for loop, you will have a problem:

[1, 1]

with this input, you will have two candidates with the same value because when you go to the second element, the second count is still 0 and you will by mistake set candidate2 to the same as candidate1.

The second trap is that when you push back the two candidates, you have to make sure that they are not the same. Because you initialize the two candidates both as 0. Then if the input gives you one majority that is 0, mistakenly you will find out the second candidate can also make the cut although it never enters the loop. That's why you have to compare them before you push in the second element.