Create 448. Find All Numbers Disappeared in an Array
The values are from 1 to n, which are a strong hint that they can be used as indices. Now, just flip them, if you can flip, remember not to flip them back. Put abs around the nums[i] because they can be flipped to negative by the previous element values.