Create 35. Search Insert Position
I wrote binary search in iterative method for the first time. If the element is found, then it will be caught in the while loop. After it exited the while loop, it must be missing in the array,  

Careful with the index here, high +  1 is the same as low, not low  -  1.