Create 240. Search a 2D Matrix II
Brilliant solution. Especially line 15 to 17. When your target is smaller, you go to left, the moment when it becomes bigger than the element, you just need to go to the next row, same column instead of the last column. This is true because if you are smaller than the previous column in the current row, you will be smaller than the previous in the next row because the values are increasing along columns.