Create 238. Product of Array Except Self
The first solution is brute force. Correct but too slow.

The second solution won't work because overflow and divide 0 elements. That reminds us that we have to visualize what we do. We all know we can divide by 0, I didn't catch that because I failed to visualize what is going on exactly.

Now comes the real deal:

We know if we know an array's left and right result, we can just multiply them together. We know the drawback of the first result is that we are wasting a lot of info. we can't store them.

The third solution is this: we go from left to right, store the multiplication of all elements left to the current element in an array, the very left element will be 1. Then we traverse from the right, store the information right to the element, multiply them together then will get the final result.

Have to say the result is pretty brilliant and I have never seen such solution before.