Create 120. Triangle
I thought I could use backtracking for this but apparently this is better of using dp. 

The problem is kind of like recursion. Thus we use the bottom up methods to solve this problem. All we can use a 2-D array to store all the necessary information for dynamic programming. But after thinking about it once we are done with row i, the information is discarded because we will never use it again. Therefore a one dimensional array Will suffice if we keep updating the information inside the array. 

The base case is the last row of the two-dimensional array. We can just use the assignments operator to assign the values off the 1D array in the first place. Then we keep updating the array itself. 

Also you can save a lot of energy by not thinking about specific indices rather than just think about how the process proceeds. For example you can just think about the process like this: you use the last row as the base case then you work your way up.

This is a very good dynamic programming problem and that is my first time actually reducing the dimensions of the dynamic programming vector.