Create 011. Container With Most Water
I didn't understand the problem clearly I thought the problem is asking you to find to adjacent lines that form the container but that's not the case. The problem is asking you to find two lines, but this two lines are not necessarily adjacent to each other. Therefore dynamic programming is not right for this kind of situation.

So now I'm thinking I can use brute force. Since is choosing to lies out of n, therefore the time complexity is n square, I can use to for loops to finish this problem. But it will exceed the time limit easily. But nonetheless let's give it a try. 

Therefore we have to search for a solution that halftime complexity better than n square. Therefore may be nlog n or n^2.

Obviously we cannot support in this case. Therefore we might be able to use some Data structures to help us know where the time complexity by sacrificing some memory space. But there's just too much information to store.

Therefore we come to our final solution. We can always use two pointers in vector. And that's our final solution. It's also some sort of greedy algorithm thinking. And that again tells us two pointers can be used things sorted or unsorted arrays.