Create 041. First Missing Positive
I was struggling to figure out this problem. But then I found out I have dealt with missing problems before. That is either sum and xor. The first solution is using sum.

Array problem O(n) can be two pointers and now we have learnt from this problem swapping. The second solution uses itself as a hash table and stores the integers as A[i] = i + 1;

One thing into a pay attention to is that we cannot use for loop in this case because we want to control the increments of i ourselves. We also don't have to care if index plus one is the elements we only need to care swapping the current elements to the right position, which is nums[i], nums[nums[i]-1].

The last part in the if statement on line 25 is very necessary because if there are two identical elements, without this part you will infinite loop because it will just swapped with itself forever for example if the input is [1, 1].