Create 069. Sqrt(x)
This problem taught me two things number one always consider overflow. One way to deal with it is to convert integer into long.

The other thing is that when you try to get the lower value of the two ends, a better way will be to return the high because the loan will be higher than the high when it breaks out of the loop.

The second solution is actually more convenient because you don't have to cast integers into long. and it also shows better understanding of final research in this case. 

The other way will be to use the Newton's method.