Create 048. Rotate Image
This is just finding patterns. If you draw a  two by three matrix, you will find out that the rotation Will yield a three by two matrix. And the previous column determines which row you are in. The previous row determines which column you're in. This is the most intuitive method.