Create 289. Game of Life
At first I I am very afraid to tackle this problem because there's just too much information. After care for thought, I found out that you just have to iterate through the board and then check if the neighbors are actually inbound. That's pretty much it.

The clever thing about this is that there are states transitions. You have to make sure that even after transitions, you can have the same current state. That's why for the life to dead transition you should use 11 because 11 mod 10 is still one. Therefore in current state, this is cell is still treated as live. During the second iteration however, you can flip all the 11s to 0. and the same thing goes to dead to live transition.