Create 073. Set Matrix Zeroes
Similar to the game of life problem. I couldn't think of a method that is in place right now, therefore I use a auxiliary Matrix to store the information of which element is going to be set to zero. 

The problem is still you don't want to use the information of the next state on your current state.

It turns out that you can use the matrix itself to store information.