Create 079. Word Search
GG walked me through the problem and it's really helpful. 

The problem is basically iterate through the matrix, see if the current element is the start of the word, then dfs the matrix to see if it can follow this current element to construct the whole word. If can, then return true, done. If not, move on to the next element. 

BFS + DFS. Backtracking happens on mark the current element as visited and if the current element is not the start of the word, unmark it and move on to the next element. 

This problem requires good knowledge of DFS. DFS can visit every element in the matrix. Every element can only go in four directions, and has to go in the direction specified by the code. When it reaches a base case, just return.

The match condition will be return true from a base case, then true from the previous direction, and then all the way return true.