Create 165. Compare Version Numbers
This problem is a good template like dfs:

while(both indices in range) is like the base case, as long as this one is good, then we keep going.

inside, the two while loops are like dfs: we keep going for one until a current condition is not met.

I spent a lot of time getting a run time error because the while loop conditions check if the indices are in range after I try to access them to see if they are '.' or not. Huge waste of time, but lesson learnt:

going in order to see if any thing is out of order.