Create 434. Number of Segments in a String
Still use the bool flag to mark continuous space chars. That is the main problem. Also use another flag to see if the first element is space or not. If is space, then you don't have to increment the counter otherwise you have to add 1 to the count otherwise you won't count the first word.