Create 033. Search in Rotated Sorted Array   Add to List QuestionEditorial Solution
http://blog.csdn.net/linhuanmars/article/details/20588511

Basically the idea is to use the fact that half of the array is still sorted after the rotation. We find out which part is still sorted by comparing the middle element to the right side: if it's smaller, then if the target fits between mid and right, look to the right; if not there look to the right.