Create 062. Unique Paths
I did it in one pass!

Same as climb ladder, the final solution is based on the sum of the previous two options since the last step are fixed. Use a 2-D array to reduce the time complexity. O(m*n) space complexity can be reduced further per the dynamic programming tutorial video.

Don't know where it's wrong with the recursion though.