Create 119. Pascal's Triangle II
very similar to the Pascal's triangle I. When it said the space limit, I knew it had to be updating the array itself. Therefore, same as I, double for loop: first one controls the size of each row and the second touches every element but the first and last one.

The trick is that the in the inner loop you are going backwards, otherwise you are using the updated values which will give you the wrong answer.