Create 179. Largest Number
I was on the right track: convert int into strings and then sort. But I didn't know there is this amazing tool called to_string so I don't have to write a function.

Second: You cannot sort the string because that will not be correct because you will break up the numbers. What you do is that you still use sort but write your own:

bool static function.

The last amazing part of the problem is tackling zero. How to deal with leading zeros. What if it's all zeros and the resulting string is empty but you still have to return a zero.

Great code, from other people.