Create 191. Number of 1 Bits
Now it seems a neat formula would be:

create a mask uint32_t mask = 1 << 31; It's important that you use uint32_t instead of int. 

Then this could be & with the number to determine whether each digit is 1 or 0.

If the return value is also uint32_t then you can use a building block 0 and gradually mask it to build the ideal number.