Create 338. Counting Bits
The first solution is based on what we have been working on today. Nothing fancy if you ask me now lol.

The second solution is more elegant: the value now can be left shifted from a previous result and the number of ones will be the same. The real question is the last digit: Is it 1 or 0.

For example, both 14 and 15 can be right shifted from 7. But 15 has one more 1 than 14 because of the last digit is 1.

So the second solution elegantly solved the problem.