Create 050. Pow(x, n)
solution 1 isO(m+n). Remember to cast both numbers now just the final result. If you just cast the final result of division of two ints, it will still be an int. 

solution 2: divide and conquer. The trick is store the n/2 part so you don't have to calculate it again. One more thing to pay attention to is abs(INT_MIN) is still INT_MIN if INT_MIN is stored in int.

spent too much time on it.