Create 300. Longest Increasing Subsequence
Nailed it after looking at lecture slides that talked about how to approach dynamic programming generally. 

From the lecture slides:

The setting is as follows. We wish to find a solution to a given problem which optimizes
some quantity Q of interest; for example, we might wish to maximize profit or minimize
cost. The algorithm works by generalizing the original problem. More specifically, it works
by creating an array of related but simpler problems, and then finding the optimal value of
Q for each of these problems; we calculate the values for the more complicated problems by
using the values already calculated for the easier problems. When we are done, the optimal
value of Q for the original problem can be easily computed from one or more values in the
array. We then use the array of values computed in order to compute a solution for the
original problem that attains this optimal value for Q. We will always present a dynamic
programming algorithm in the following 4 steps.