Create 91. Decode Ways
Name of the game is every one letter and every two letters. If the one letter is valid, then attach one fixed letter will be the same as dp(n-1). Same thing goes to two letters. What makes the code so elegant is that it considers two cases separately and gets rid of a lot of consideration of situations.