Create 394. Decode String
Still same mistake on multiple situations: put the most important one first: line 22.

The problem does a good job listing possible situations. There is little algorithm left except you have to figure out you have to use stack for this kind of problems. 

Note that the stack in this case is always to work pool. All things have to go in the stack, even if you just pull things out of it.  

Couple days before it's such a daunting task, now it is alright.