Create 325. Maximum Size Subarray Sum Equals k
The first option is to use double for loop. Maybe has the right answer, but TLE.

Now whether we use recursion or dp, there is this question: how do you get the scenarios where the subarray doesn't start from the first element.

The answer is using hash table. Then you can subtract indices to get whatever part you want. Basically two parts: part of the array can get you to sum - k, where k = 0 in this case, and part that can get you to sum. Then the contiguous array will be the difference between the indices.

Also note when you update map[sum]. You want the indices to be as small as possible so when you subtract the indices you can have a larger subarray.

This is very similar to problem 325.