Create 270. Closest Binary Search Tree Value
First up is to use the in-order traversal property of BST. O(n).

Then I looked at the tag and realize it's asking me to search in a binary search tree. Of course, there is a O(logn) solution.

used the iterative solution for the first time.

Then the recursion method.