Create 172. Factorial Trailing Zeroes
First thought: calculate the product and then calculate the trailing zeroes. 

Then the only time trailing zeroes can happen is when there is 2*5. Since there is going to be a lot of 2s in the n!, the final result is limited by the number of 5s or 25s, or 125s. 

Therefore you can arrive at the final solutions.

The first one has overflow and the second one use division to avoid overflow.