Create 208. Implement Trie (Prefix Tree)
My first thought, which is wrong. It uses a flag to separate words and prefixes, which is also a key idea in the right implementation of the final solution. 


This problem also gives interesting point: I was hesitant to use treenode structure in the first place because I was thinking it's impossible to have 26 pointers like pointers in binary tree. But later the use of map just illustrates pointers are just addresses. 

After that it's good.