Create 311. Sparse Matrix Multiplication
In perfect hindsight, this problem is straightforward. The brute force way would be the normal matrix multiplication. 

But since the matrix is sparse, then a better way to do it will be use two hash tables. One to record the non-zero rows in A, the other to record the non-zero column in B. Then only multiply the contents in the two hash tables to come up with the final result. This is because we record the row and column indices and the final location of the inner products are determined by the two indices. To calculate the inner products we use the standard library tool.

I spend maybe an hour on it because I got confused about the numbers of the rows and columns. But now it becomes more clear to me.