Create 280. Wiggle Sort
I first thought it's like linked list problem where you can have the first part of the SORTED array and interlace it with the second part of the array. But that will involve extra memory.

The problem is from other people and it's quite sneaky. It just corrects violations at the even or odd positions by swapping. Why is this algorithm correct? This is again finding patterns.