Create 324. Wiggle Sort II
The first solution is based on Wiggle sort 1 with minor correction on how and when to swap. But the problem lies in this 

1 3 1 1.

if you swap 1 and 1, it still has a problem. Then you might think that if you use while to find the first swappable element, which is 3, then the natural question will arise: you will break the pattern at 3.

So it's not that simple. 

One solution will still be break the array in to two parts, and then you will interlace them together.