Create 314. Binary Tree Vertical Order Traversal
My first solution uses dfs but it fails. 

It is quite concise and elegant. I got the most of the ideas down:

I have the level to indicate whether it should go left or not. I used the map to record all the possible columns. 

I used height in a pair to figure out the order of the elements in the same column.

I used the dfs properties to record elements on the same row and column.

But the height is just a killer. So much of the work involves sorting out the height, it's just too much.

BFS will be a much better one.

Still BFS can go from left to right to take care of same column same row problem. It can mark column number by marking nodes in the queue. It doesn't care about height because it's BFS. It doesn't need a helper function because it's iterative.