Create 498. Diagonal Traverse
I  think this problem is a confidence booster. I was afraid of this kind of question, but now I at least solved one of this pattern finding questions. 

The problem obviously requires visiting every element. So using count as the termination condition is a good idea. I further break the problem into two conditions: the trend is going up and going down.

If you study the example closely, you will find out that there are two kinds of switches, reflected in the code for both up and down. Namely how to move the pointers to the first element in the next direction.