Create 286. Walls and Gates
My correct but really slow solution. In the main loop, only 0 has the real power to change thing, -1 never changes and INT_MAX just reacts. Therefore only deals with 0 elements.

In the bfs part, go four directions after check they are inbound and they are not -1 or 0 because they are supposed to change but 0 and -1 don't. Also add a map to prevent backtracking. 

dfs on the other hand is much more concise. When you cannot change the status quo for the better just stop.

My first time to use BFS and DFS both to solve a problem.