Create 243. Shortest Word Distance
This problem is not hard, but it asks a lot of questions we are used to ignore. 

For example, are the two words really in the array. Are they guaranteed not to be the same? We are so used to working on the problems right away and not to ask any question.