Create 256. Paint House
My first method is like this:

The dp array has two parameters: one is the cost, the other is the color chosen for the previous house. The current cost will be the  previous cost + the min of the current colors. But guess what? It's wrong because it's not dynamic programming. It's greedy algorithm because it's choosing the min cost for the current colors every single time.

Consider this example:

1  2     3     4                1   2      3    4

2 100 100 100             4  101 101 101 

1 1     1     1                 102 5   5    5 

You will choose 1 in the first row, but then you regret that but you cannot backtrack because you didn't store the necessary information.

Now I think I have a better understanding of dp. I did it because I think of the dungeon path problem. You are supposed to memorize all the possible values, and then choose from the previous values, like what I did in the second solution.