Create 245. Shortest Word Distance III
To use hash table for this kind of problems is not hard. But I was struggling with this problem:

how do you find the min difference between two sorted arrays. After all, this is what the problem comes down to.

Now this array can have two forms:

two different vectors(word1 and word2 are different), or two identical vectors(two words are the same). 

Pay attention to the while loop. When they are different, we try to update the diff. After that we adjust the indices. Smart.