Create 253. Meeting Rooms II
I actually got most of the problem down. You have to sort the array based on start time. Then it's a matter of seeing end time.

My solution based on the following logic: you want the start time to be as close to the end time as possible to minimize the number of rooms needed. Then I used a set to store the end time. Then I traverse the set from the biggest end time to the smallest end time. When there is an end time smaller than the start time, I will update that room's end time by deleting the old end time and add the new end time. I also used multiset because there can be two meetings end at the same time. It doesn't work.

So the solution is more dynamic. It uses a set to store the end time still. Now where there is a new request for room, it goes through the set: is this room finished by determining if the end time is smaller than the start time. If so, take it out of the set. Do this until you cannot find a vacant room, then you insert this end time. 

The beauty is you keep recording the size of the set, and return the max. This makes sense because the most hectic hours determined how many rooms you need at a minimum.