Create 246. Strobogrammatic Number
This problem uses two pointers and is based on pattern finding.

We know three very reliable numbers: 0, 1, 8. And 6, 9.

I did some pattern finding and thought about using hash table as the tag suggests. But I found out it's just two pointers. The thing with hash table is that it only has counts instead of the positions, which matters greatly in this case. 818 is ok but 881 is not. Also I blend 6 and 9 with 0 and 1. If the two are put symmetrically then they are good. Otherwise it's not.

It's just pattern finding. Like palindrome.