Create 127. Word Ladder
http://blog.csdn.net/qq508618087/article/details/51344102

BFS on graph. When to stop; When to mark, how to mark vistied; how to do BFS level by level. How to modify a string to speed up.

In perfect hindsight, this problem is not hard, and the above link has a solution that is even more concise.

BFS, pick a starting point, push all possible but not visited nodes into the queue until there is nothing left in the queue. Also the classic two containers swapping to ensure the right BFS.

This problem is indeed good.