Create 316. Remove Duplicate Letters
The solution is just amazing and easy to understand.

Also for duplicates, you have to think about set and vector of chars as a hash table.

You are using two hash tables.

Go through the string, If you have this char in the final result, then continue, because adding it to the final result will be duplicates. Then you need a counter, if I forgo this one, do I have another left in the original longer string.

Then if the current last element in the result string is in the wrong spot and there are more in the original string later, you kick them out until if you add the char to the result string, you won't wreck the order of the result string. 

You also have to make sure that you are keeping track of the counter and the visited hash table, i.e., when you add chars to the final string you change the corresponding bool and also the same when you delete char from the final string.

We are using the string as a stack here. But again, it shows that when it comes to ordering, stack is great.