Create 115. Distinct Subsequences
The first exercise is a very good practice of recursion and gives us a preview into what the dp solution could be. 

Now the base case is if the two are the same, return 1. If s is shorter than t, then return 0.

Onto the recursion case, we know there must be a recursion + x part somewhere in the function. Because the final result can be more than 1. Actually we have done something like that before.

Now let's see if we can work out the dp solution.

http://blog.csdn.net/qq508618087/article/details/50108715