Create 261_graph_valid_tree.cpp
Almost have it done in one pass. But forgot to push in second line into the map when building the adjacency list.

A tree is an undirected, acyclic, connected graph.

Therefore, we check if it's one connected component and then we check if it has cycle or not. Both are done in DFS.

On a second thought, when you check acyclic, you are performing one dfs. That is exactly what connected is trying to do. So now all you have to do is to check if all the nodes have been visited or not.